requirements.txt
asgiref==3.3.1
Django==3.1.5
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
dj-database-url==0.5.0
django-braces==1.14.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.5
six==1.15.0
whitenoise==5.2.0